massive cleanup changes routes adds styling
