+++ src/components/container/NotFound.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 
 const NotFound = () => {
   return (
-    <div className="404-error">
+    <div className="404-error block">
       <h2>Not Found</h2>
       <img src="https://i.pinimg.com/736x/5c/97/c8/5c97c8c2afa2cde40aa6330266389d90--anime-meme-funny-dogs.jpg" alt="Dog getting hit with frisbee" />
     </div>
