+++ src/components/presentational/Header.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
+import { NavLink as Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
 class Header extends Component {
